HomeWorks v2.2.1 Fix: clean code

+++ src/com/company/Tasks/BasicOfPrograming/bitCharCounter.java
@@ -15,7 +15,7 @@ public class bitCharCounter {
         //conversion
         stringToChar = input.charAt(0);
         //conversion
-        charToInt = (int) stringToChar;
+        charToInt = stringToChar;
         //conversion
         intToBinaryString = Integer.toBinaryString(charToInt);
